+++ kernel-hal-unix/src/lib.rs
@@ -23,9 +23,6 @@ use {
     tempfile::tempdir,
 };
 
-mod fb;
-
-pub use fb::*;
 pub use kernel_hal::defs::*;
 use kernel_hal::vdso::*;
 pub use kernel_hal::*;
@@ -439,6 +436,80 @@ pub fn init_framebuffer() {
         return;
     }
 
+    #[repr(C)]
+    #[derive(Debug, Default)]
+    struct FbFixScreeninfo {
+        id: [u8; 16],
+        smem_start: u64,
+        smem_len: u32,
+        type_: u32,
+        type_aux: u32,
+        visual: u32,
+        xpanstep: u16,
+        ypanstep: u16,
+        ywrapstep: u16,
+        line_length: u32,
+        mmio_start: u64,
+        mmio_len: u32,
+        accel: u32,
+        capabilities: u16,
+        reserved: [u16; 2],
+    }
+
+    impl FbFixScreeninfo {
+        pub fn size(&self) -> u32 {
+            self.smem_len
+        }
+    }
+
+    #[repr(C)]
+    #[derive(Debug, Default)]
+    struct FbVarScreeninfo {
+        xres: u32,
+        yres: u32,
+        xres_virtual: u32,
+        yres_virtual: u32,
+        xoffset: u32,
+        yoffset: u32,
+        bits_per_pixel: u32,
+        grayscale: u32,
+        red: FbBitfield,
+        green: FbBitfield,
+        blue: FbBitfield,
+        transp: FbBitfield,
+        nonstd: u32,
+        activate: u32,
+        height: u32,
+        width: u32,
+        accel_flags: u32,
+        pixclock: u32,
+        left_margin: u32,
+        right_margin: u32,
+        upper_margin: u32,
+        lower_margin: u32,
+        hsync_len: u32,
+        vsync_len: u32,
+        sync: u32,
+        vmode: u32,
+        rotate: u32,
+        colorspace: u32,
+        reserved: [u32; 4],
+    }
+
+    impl FbVarScreeninfo {
+        pub fn resolution(&self) -> (u32, u32) {
+            (self.xres, self.yres)
+        }
+    }
+
+    #[repr(C)]
+    #[derive(Debug, Default)]
+    pub struct FbBitfield {
+        offset: u32,
+        length: u32,
+        msb_right: u32,
+    }
+
     let mut vinfo = FbVarScreeninfo::default();
     if unsafe { libc::ioctl(fbfd, FBIOGET_VSCREENINFO, &mut vinfo) } < 0 {
         return;
@@ -489,8 +560,7 @@ type KBDCallbackFn = dyn Fn(u16, i32) + Send + Sync;
 
 lazy_static! {
     static ref MOUSE_CALLBACK: Mutex<Vec<Box<MouseCallbackFn>>> = Mutex::new(Vec::new());
-    static ref KBD_CALLBACK: Mutex<Vec<Box<KBDCallbackFn>>> =
-        Mutex::new(Vec::new());
+    static ref KBD_CALLBACK: Mutex<Vec<Box<KBDCallbackFn>>> = Mutex::new(Vec::new());
 }
 
 #[export_name = "hal_mice_set_callback"]
